Input: synaptics-rmi4: Use generic interrupt handling
Currently the RMI4 driver expects the device to have a GPIO and manages
the that GPIO internally. However, this duplicates functionality which
could be handled by more generic interrupt handling code. Also, some
RMI devices will not have a GPIO or it won't be accessible to the rmi4
driver. This patch removes the GPIO code and instead gets the irq passed
up from the underlying transport (ie i2c-core).

Signed-off-by: Andrew Duggan <aduggan@synaptics.com>

Conflicts:
	drivers/input/rmi4/rmi_driver.c
	drivers/input/rmi4/rmi_driver.h